requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
Django==4.2.16
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
redis==5.1.1
sqlparse==0.5.1
typing-extensions==4.12.2
